Improvements and fixes for configs
